+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## UNRELEASED
+
+- Fixed incorrectly raised exception when using custom scalar as query argument type.
+
+
 ## 0.2.0 (2023-02-02)
 
 - Added `remote_schema_url` and `remote_schema_headers` settings to support reading remote schemas.